Trim unnecessary tests
